# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MagicAI  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/general/favorite-openai": {
            "get": {
                "tags": [
                    "General (Helpers)"
                ],
                "summary": "Get last 6 favorite OpenAI items",
                "description": "Get the last 6 favorite OpenAI items for the authenticated user.",
                "operationId": "openAIFavoritesList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Post Title Generator"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Get captivating post titles instantly with our title generator. Boost engagement and save time."
                                            },
                                            "slug": {
                                                "type": "string",
                                                "example": "post_title_generator"
                                            },
                                            "active": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "questions": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "image": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "premium": {
                                                "type": "integer",
                                                "example": 0
                                            },
                                            "type": {
                                                "type": "string",
                                                "example": "text"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "example": "2023-03-11T08:26:49.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "example": "2023-03-11T08:26:49.000000Z"
                                            },
                                            "prompt": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "custom_template": {
                                                "type": "integer",
                                                "example": 0
                                            },
                                            "tone_of_voice": {
                                                "type": "integer",
                                                "example": 0
                                            },
                                            "color": {
                                                "type": "string",
                                                "example": "#A3D6C2"
                                            },
                                            "filters": {
                                                "type": "string",
                                                "example": "blog"
                                            },
                                            "package": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "pivot": {
                                                "properties": {
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "openai_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized: User not authenticated"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
        "/api/affiliates/": {
            "get": {
                "tags": [
                    "Affiliates"
                ],
                "summary": "Get affiliate totals of current user",
                "description": "Get affiliate earnings and withdrawal totals of current user.",
                "operationId": "affiliates",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/affiliates/withdrawals": {
            "get": {
                "tags": [
                    "Affiliates"
                ],
                "summary": "Get withdrawals of current user",
                "description": "Get affiliate withdrawals list of current user.",
                "operationId": "withdrawals",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
        "/api/support/user/{ticket_id}": {
            "get": {
                "tags": [
                    "Support"
                ],
                "summary": "Gets information of the user which sent the request",
                "description": "Gets information of the user which sent the request. Use ticket ids like QZDNGSIFPH not integers.",
                "operationId": "ticketUser",
                "parameters": [
                    {
                        "name": "ticket_id",
                        "in": "path",
                        "description": "Ticket ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "QZDNGSIFPH"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Ticket / User Not Found"
                    },
                    "412": {
                        "description": "Precondition Failed"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://127.0.0.1:8000/oauth/authorize",
                        "tokenUrl": "http://127.0.0.1:8000/oauth/token",
                        "refreshUrl": "http://127.0.0.1:8000/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}